+++ packages/authentication/modules/user/getUserAsync/types.ts
@@ -1,0 +1,5 @@
+import type { ServerSideRenderingOption } from '@baseapp-frontend/utils'
+
+export interface GetUserAsyncOptions extends ServerSideRenderingOption {
+  key?: string
+}