+++ packages/authentication/modules/user/useJWTUser/types.ts
@@ -1,13 +1,13 @@
-import { ServerSideRenderingOption } from '@baseapp-frontend/utils'
+import type { ServerSideRenderingOption } from '@baseapp-frontend/utils'
 
 import UserApi from '../../../services/user'
-import { CustomCookieNames } from '../../../types/auth'
-import { CustomUseQueryOptions } from '../../../types/react-query'
+import type { CustomJWTKeyNames } from '../../../types/auth'
+import type { CustomUseQueryOptions } from '../../../types/react-query'
 
 type ApiClass = Pick<typeof UserApi, 'getUser'>
 
 export interface UseJWTUserOptions<TUser>
-  extends CustomCookieNames,
+  extends CustomJWTKeyNames,
     ServerSideRenderingOption,
     CustomUseQueryOptions<TUser, unknown, TUser> {
   ApiClass?: ApiClass