+++ packages/utils/constants/cookie.ts
@@ -1,3 +1,9 @@
+/**
+ * @deprecated Use ACCESS_KEY_NAME instead
+ */
 export const ACCESS_COOKIE_NAME = 'Authorization'
+/**
+ * @deprecated Use REFRESH_KEY_NAME instead
+ */
 export const REFRESH_COOKIE_NAME = 'Refresh'
 export const LANGUAGE_COOKIE_NAME = 'Language'