+++ packages/utils/functions/cookie/index.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import ClientCookies, { CookieAttributes } from 'js-cookie'
 
 import { GetCookieOptions } from './types'
@@ -17,6 +18,21 @@ export const getCookie = <T>(
   let cookie
   if (typeof window === typeof undefined && !noSSR) {
     const { cookies: serverCookies } = require('next/headers')
+    cookie = serverCookies().get(key)?.value
+  } else {
+    cookie = ClientCookies.get(key)
+  }
+  return parseJSON ? parseCookie<T>(cookie) : (cookie as T)
+}
+
+export const getCookieAsync = async <T>(
+  key: string,
+  { noSSR = false, parseJSON = true }: GetCookieOptions = {},
+) => {
+  let cookie
+  if (typeof window === typeof undefined && !noSSR) {
+    // @ts-ignore
+    const { cookies: serverCookies } = await import('next/headers')
     cookie = serverCookies().get(key)?.value
   } else {
     cookie = ClientCookies.get(key)