+++ packages/utils/functions/token/refreshAccessToken/__tests__/refreshAccessToken.test.ts
@@ -1,9 +1,10 @@
 import Cookies from 'js-cookie'
 
 import { refreshAccessToken } from '..'
-import { ACCESS_COOKIE_NAME, REFRESH_COOKIE_NAME } from '../../../../constants/cookie'
+import { ACCESS_KEY_NAME, REFRESH_KEY_NAME } from '../../../../constants/jwt'
 import { getAccessToken } from '../../getAccessToken'
 import { getToken } from '../../getToken'
+import { setTokenAsync } from '../../setTokenAsync'
 
 jest.mock('../../getAccessToken')
 jest.mock('../../getToken')
@@ -11,6 +12,8 @@ jest.mock('js-cookie', () => ({
   set: jest.fn(),
   remove: jest.fn(),
 }))
+
+jest.mock('../../setTokenAsync')
 
 describe('refreshAccessToken', () => {
   beforeEach(() => {
@@ -26,18 +29,16 @@ describe('refreshAccessToken', () => {
     const getAccessTokenMock = getAccessToken as jest.Mock
     getAccessTokenMock.mockResolvedValue(newAccessToken)
 
-    const result = await refreshAccessToken()
+    await refreshAccessToken()
 
-    // Verify
-    expect(getToken).toHaveBeenCalledWith(REFRESH_COOKIE_NAME)
+    expect(getToken).toHaveBeenCalledWith(REFRESH_KEY_NAME)
     expect(getAccessToken).toHaveBeenCalledWith(refreshToken)
-    expect(Cookies.set).toHaveBeenCalledWith(ACCESS_COOKIE_NAME, newAccessToken, {
+    expect(setTokenAsync).toHaveBeenCalledWith(ACCESS_KEY_NAME, newAccessToken, {
       secure: false,
     })
-    expect(result).toBe(newAccessToken)
   })
 
-  it('should remove cookies if refreshing the access token fails', async () => {
+  it('should remove tokens if refreshing the access token fails', async () => {
     const refreshToken = 'valid-refresh-token'
     const getTokenMock = getToken as jest.Mock
     getTokenMock.mockReturnValue(refreshToken)
@@ -46,19 +47,19 @@ describe('refreshAccessToken', () => {
     getAccessTokenMock.mockRejectedValue(new Error('Failed to refresh token'))
 
     await expect(refreshAccessToken()).rejects.toThrow('Failed to refresh token')
-    expect(getToken).toHaveBeenCalledWith(REFRESH_COOKIE_NAME)
+    expect(getToken).toHaveBeenCalledWith(REFRESH_KEY_NAME)
     expect(getAccessToken).toHaveBeenCalledWith(refreshToken)
-    expect(Cookies.remove).toHaveBeenCalledWith(ACCESS_COOKIE_NAME)
-    expect(Cookies.remove).toHaveBeenCalledWith(REFRESH_COOKIE_NAME)
+    expect(Cookies.remove).toHaveBeenCalledWith(ACCESS_KEY_NAME)
+    expect(Cookies.remove).toHaveBeenCalledWith(REFRESH_KEY_NAME)
   })
 
-  it('should remove cookies if no refresh token is available', async () => {
+  it('should remove tokens if no refresh token is available', async () => {
     const getTokenMock = getToken as jest.Mock
     getTokenMock.mockReturnValue('')
 
     await expect(refreshAccessToken()).rejects.toThrow()
-    expect(getToken).toHaveBeenCalledWith(REFRESH_COOKIE_NAME)
-    expect(Cookies.remove).toHaveBeenCalledWith(ACCESS_COOKIE_NAME)
-    expect(Cookies.remove).toHaveBeenCalledWith(REFRESH_COOKIE_NAME)
+    expect(getToken).toHaveBeenCalledWith(REFRESH_KEY_NAME)
+    expect(Cookies.remove).toHaveBeenCalledWith(ACCESS_KEY_NAME)
+    expect(Cookies.remove).toHaveBeenCalledWith(REFRESH_KEY_NAME)
   })
 })